`cursor` should not be a parameter the end user needs to think about
At the moment, the end user has to deal with pagination - the first
request returns a `cursor`, and the user has to pass in the cursor to
the endpoint to get the next page. When the `cursor` is `None`, the
final page has been reached.

This is unnecessary complexity; if a `limit` is passed, the SDK should
just return the first `limit` records, and if a `limit` is not passed,
the SDK should do all of the pagination itself to get the full results
before returning to the end user.

The downside of this is that there can be a delay before the client can
start processing the first record, but this shouldn't be an issue as for
speed critical endpoints, the user should be using websockets.